+++ commands/ex/ExW.cpp
@@ -8,7 +8,7 @@ VM::ExW::ExW() {
 void VM::ExW::executeFor(VM::VMModel &model, const std::string &inputSequence) {
     WindowState *state = model.getActiveState();
     std::string newFileName;
-    for (int i = 2; i < inputSequence.length(); ++i) {
+    for (size_t i = 2; i < inputSequence.length(); ++i) {
         if (inputSequence.at(i) != ' ') {
             newFileName = inputSequence.substr(i);
             break;
